[spinel] add support for spinel interface id
- This feature allows the RCP to support multiple host stacks on different PANs by making use of the spinel Interface ID.
- Some minor cleanup.
- [spinel] Revert RCP recovery changes.
- Bug fixes...
 It fixes following bugs.
 1) In ncp_base.cpp, push the CurrentCommandIID to queue to be used by PandingEnergyScanHandler.
 2) In sub_mac.cpp, check for kStateDelayBeforeRetx when deciding if the device is transmitting or scanning. When there is an Ack Timeout, the RCP by default will try to re-transmit the failed packet, but is configured to give a random backoff during this period. When it does it will be in kStateDelayBeforeRetx state. If not careful, during the backoff period, RCP might accept another packet to transmit and face issues when sub_mac begins the transmission again.
 3) In radio_spinel_impl.hpp, when the RCP recovery begins, the SPINEL tries to move to Rx state, switch channels to initialise the RCP. But these should not be performed when running as Multiprotocol RCP. So, if the RCP throws an error, ignore it.
 4) Fix git workflow errors.
- [spinel] Make OPENTHREAD_SPINEL_CONFIG_RCP_TX_WAIT_TIME_SECS option configurable during compile time.
Create unit tests for NcpBase pending command queue
The multipan rcp configuration requires the use of a queue for spinel commands which require the use of the radio when it is in a busy state. A pending command queue was added to service one or more hosts issuing requests to the secondary.

This test suite introduces tests to cover the conditions where commands need to be queued and handled later by controlling the order of spinel frames and callbacks and validating the spinel status when commands are issued.