Initialize key instance offset in KeyRefManager constructor
Contention can arise during the instance initialization process in which the KeyRefManager initializes
the reference offset to 0 before the Instance class updates the offset with an instance specific value.
Before the instances can update the offset, later instances will perform key export and desctruction
operations using the reference addresses of earlier instances, which will cause keys to not be
found temporarily.

The effect of this situation is that an existing network key from the active dataset can not be
found for any instance on device start up, only when the key is set after initialization. When
a device restarts and tries to reconnect to a network, the device will be able to connect and join,
but the network key update will restart the device's frame counter, which will cause security errors
for secured message exchanges, such as an address solicit.

This commit moves the KeyRefManager offset update for multiple instances to the KeyRefManager constructor
rather than the instance constructor to avoid this initialization contention.